requirements.txt
peewee~=3.16.3
requests~=2.31.0
python-dotenv~=1.0.0
pyTelegramBotAPI~=4.13.0
PyMySQL==1.1.0
cryptography==41.0.4